requirements.txt
streamlit
feedparser
pyzbar
Pillow
python-dotenv
google-generativeai


